requirements.txt
-i https://pypi.org/simple
allure-pytest==2.12.0
allure-python-commons==2.12.0
appdirs==1.4.4
assertpy==1.1
astroid==2.12.13
async-generator==1.10 ; python_version >= '3.5'
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
colorama==0.4.6
configparser==5.3.0
crayons==0.4.0
dill==0.3.6 ; python_version >= '3.11'
distlib==0.3.6
enums==0.0.2
fake-useragent==1.1.1
faker==15.3.4
filelock==3.8.2
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.29
h11==0.14.0 ; python_version >= '3.7'
idna==3.4
iniconfig==1.1.1
isort==5.11.3
lazy-object-proxy==1.8.0
mccabe==0.7.0
outcome==1.2.0 ; python_version >= '3.7'
packaging==22.0 ; python_version >= '3.7'
page-objects==1.1.0
pipenv==2022.12.17
platformdirs==2.6.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
py==1.11.0
pylint==2.15.9
pyparsing==3.0.9
pysocks==1.7.1
pytest==7.2.0
pytest-ordering==0.6
python-dateutil==2.8.2
python-dotenv==0.21.0 ; python_version >= '3.7'
requests==2.28.1
selenium==4.7.2
setuptools==65.6.3 ; python_version >= '3.7'
six==1.16.0
smmap==5.0.0 ; python_version >= '3.6'
sniffio==1.3.0 ; python_version >= '3.7'
sortedcontainers==2.4.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.11.6 ; python_version >= '3.6'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
trio==0.22.0 ; python_version >= '3.7'
trio-websocket==0.9.2 ; python_version >= '3.5'
urllib3==1.26.13
virtualenv==20.17.1
virtualenv-clone==0.5.7
webdriver-manager==3.8.5
wrapt==1.14.1
wsproto==1.2.0 ; python_full_version >= '3.7.0'
xlrd==2.0.1
